requirements.txt
langchain-community
pypdf
langchain-huggingface
langchain-chroma
sentence-transformers
langchain-groq
python-dotenv